+++ scripts/CTestConfig.cmake
@@ -4,4 +4,3 @@ set(CTEST_DROP_METHOD "http")
 set(CTEST_DROP_SITE "my.cdash.org")
 set(CTEST_DROP_LOCATION "/submit.php?project=bitsery")
 set(CTEST_DROP_SITE_CDASH TRUE)
-