+++ tests/serialization_ext_compact_value.cpp
@@ -26,7 +26,6 @@
 #include "serialization_test_utils.h"
 
 #include <bitsery/traits/array.h>
-#include <iostream>
 #include <bitset>
 #include <chrono>